+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.3.2] - 2026-04-29
+
 - Fix Nemesis/Heracles and Icarus/Artemis being able to spawn in the same room
 - Increase cooldown between two field NPC encounters from 6 to 8 rooms
 
@@ -25,7 +27,8 @@
 
 - First version of the mod
 
-[unreleased]: https://github.com/adi1998/ExtendedNPCEncounters/compare/0.3.1...HEAD
+[unreleased]: https://github.com/adi1998/ExtendedNPCEncounters/compare/0.3.2...HEAD
+[0.3.2]: https://github.com/adi1998/ExtendedNPCEncounters/compare/0.3.1...0.3.2
 [0.3.1]: https://github.com/adi1998/ExtendedNPCEncounters/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/adi1998/ExtendedNPCEncounters/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/adi1998/ExtendedNPCEncounters/compare/0.2.0...0.2.1